Fix power production buildings showing in Power tab.
Power producers now have a PowerComp that is a subtype of CompPowerTrader - the vanilla def.HasComp only checks exact matches, and did not catch them.

Added HasCompOrChildCompOf() extension for defs that uses IsAssignableFrom to do the type check. Not entirely foolproof (e.g. int can be assigned from float but is not a parent of float), but within the constraints of this problem it should be fine.