cheesy 'fix' for #57
Implementing different counting schemes for corpses and other products would be a pita. Instead, let's avoid the problem by limiting the filter to foodstuffs only.

Note that on existing games/jobs, if a non-food item was previously set to be counted, it will still be set - but not show up on the GUI anymore. Create a new job to fix that.

resolves #57